requirements.txt
# Telegram Bot
aiogram==3.13.1
aiohttp==3.10.10

# Environment variables
python-dotenv==1.0.1

# Database
aiosqlite==0.20.0
asyncpg==0.29.0
supabase==2.10.0

# AI Integration
openai==1.55.3

# Date handling
python-dateutil==2.9.0

# Utility
pydantic==2.9.2

# HTTP requests (для scripts/setup_webhook.py)
requests==2.31.0

# Flask для веб-приложения
Flask==3.0.3
